webapp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
langchain==0.0.178  # Adjust version based on your actual usage
langchain_openai==0.0.178  # Adjust version based on your actual usage
langchain_experimental==0.0.178  # Adjust version based on your actual usage
googlemaps==4.6.0
resend==0.1.4
requests==2.28.2
PyJWT==2.8.0
python-dotenv==0.21.0